+++ src-docs/src/views/scroll/scroll_y.tsx
@@ -37,7 +37,6 @@ export default () => {
             <ScrollContent />
           </div>
         }
-        snippetLanguage="tsx"
         snippet={`<div
   tabIndex={0}
   role="region"
@@ -82,7 +81,6 @@ export default () => {
               <ScrollContent />
             </div>
           }
-          snippetLanguage="tsx"
           snippet={
             `<div
   tabIndex={0}