+++ src/components/beacon/beacon.tsx
@@ -10,6 +10,8 @@ import React, { FunctionComponent, HTMLAttributes } from 'react';
 import { CommonProps } from '../common';
 import classNames from 'classnames';
 
+import { euiBeaconStyles } from './beacon.styles';
+
 export type EuiBeaconProps = Omit<HTMLAttributes<HTMLDivElement>, 'children'> &
   CommonProps & {
     /**
@@ -25,12 +27,16 @@ export const EuiBeacon: FunctionComponent<EuiBeaconProps> = ({
   ...rest
 }) => {
   const classes = classNames('euiBeacon', className);
+  const styles = euiBeaconStyles();
+  const cssStyles = [styles.euiBeacon];
 
-  const styles = {
+  const beaconStyle = {
     ...style,
     height: size,
     width: size,
   };
 
-  return <div className={classes} style={styles} {...rest} />;
+  return (
+    <div className={classes} css={cssStyles} style={beaconStyle} {...rest} />
+  );
 };