+++ src/components/title/title.test.tsx
@@ -9,6 +9,7 @@
 import React from 'react';
 import { render } from 'enzyme';
 import { requiredProps } from '../../test';
+import { shouldRenderCustomStyles } from '../../test/internal';
 
 import { EuiTitle } from './title';
 
@@ -22,6 +23,12 @@ describe('EuiTitle', () => {
 
     expect(component).toMatchSnapshot();
   });
+
+  shouldRenderCustomStyles(
+    <EuiTitle>
+      <h1>Title</h1>
+    </EuiTitle>
+  );
 
   test('renders children element className', () => {
     const component = render(