+++ src/themes/amsterdam/global_styling/variables/_states.ts
@@ -8,23 +8,24 @@
 
 import { computed } from '../../../../services/theme/utils';
 import { transparentize } from '../../../../services/color';
-import {
-  focus,
-  _EuiThemeFocus,
-} from '../../../../global_styling/variables/_states';
+import { _EuiThemeFocus } from '../../../../global_styling/variables/states';
+import { sizeToPixel } from '../../../../global_styling/functions/size';
 
 /**
  * NOTE: These were quick conversions of their Sass counterparts.
  *       They have yet to be used/tested.
  */
 
-export const focus_ams: _EuiThemeFocus = {
-  ...focus,
+export const focus: _EuiThemeFocus = {
   color: 'currentColor',
   transparency: { LIGHT: 0.9, DARK: 0.7 },
   backgroundColor: computed(({ colors, focus }) =>
     transparentize(colors.primary, focus!.transparency)
   ),
+
+  // Sizing
+  widthLarge: computed(sizeToPixel(0.25)),
+  width: computed(sizeToPixel(0.125)),
 
   // Outline
   outline: {